+++ application/src/main/java/com/hanwha/urlshortener/db/repository/UrlRepository.java
@@ -5,7 +5,7 @@ import com.hanwha.urlshortener.db.entity.Url;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UrlRepository extends JpaRepository<Url, Integer> {
+public interface UrlRepository extends JpaRepository<Url, Long> {
 
   Optional<Url> findByOriginalURL(String shortURL);
 }
